# contracts/openapi/pie-control-plane-v1.yaml
openapi: 3.1.2
info:
  title: Pie Control Plane API
  version: 0.1.0
  description: Initial executable contract for instance discovery, session, organizations, projects, and work items.
servers:
  - url: https://api.pielab.ai
security:
  - PieOidc: []
...
paths:
...
  /v1/organizations/{organizationId}/notifications/preferences:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [Notifications]
      operationId: getNotificationPreferences
      responses:
        '200':
          description: The caller's desktop, DND, and channel notification preferences.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/notification-preferences.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
    put:
      tags: [Notifications]
      operationId: updateNotificationPreferences
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/resources/notification-preferences-update.v1.schema.json
      responses:
        '200':
          description: Updated preferences.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/notification-preferences.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/channels/{channelId}/notification-level:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    put:
      tags: [Collaboration]
      operationId: setChannelNotificationLevel
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [level]
              properties:
                level:
                  enum: [all, mentions, none]
              additionalProperties: false
      responses:
        '204':
          description: Channel notification level updated.
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/organizations/{organizationId}/meeting-participants/{participantId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: participantId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    patch:
      tags: [Meetings]
      operationId: updateMeetingParticipantRole
      description: Changes a non-host participant role and forces a fresh capability-scoped token.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: ../schemas/resources/meeting-participant-role-update.v1.schema.json
      responses:
        '200':
          description: Participant with the updated role.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-participant.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    PieOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.pielab.ai/realms/pie/.well-known/openid-configuration
      description: The actual issuer is discovered from /.well-known/pie for each instance.
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        $ref: ../schemas/common/opaque-id.v1.schema.json
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
...
  responses:
    Problem:
      description: RFC 9457 problem with a stable Pie error code.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/common/problem-details.v1.schema.json